Execute the callable of a stageRunnerNode.

Execute the callable of a stageRunnerNode.

stageRunnerNode_run(..., .cached_env = NULL, .callable = self$callable)

Arguments

...
additional arguments to the callable. This allows, stagerunner stages to be uniformly parametrized (for example, if all stages should have a verbose parameter.

.cached_env
An internal helper that passes the cached environment to be used for storing the results of this execution.

.callable
Another internal helper used for some recursive metaprogramming.

Value

TRUE if the execution was successful, or an error otherwise.