Overlaying a stageRunner object is taking another stageRunner object with similar stage names and adding the latter's stages as terminal stages to the former (for example, to support tests).

Overlaying a stageRunner object is taking another stageRunner object with similar stage names and adding the latter's stages as terminal stages to the former (for example, to support tests).

stageRunner_overlay(other_runner, label = NULL, flat = FALSE)

Arguments

other_runner
stageRunner. Another stageRunner from which to overlay.

label
character. The label for the overlayed stageRunner. This refers to the name the former will get wrapped with when appended to the stages of the current stageRunner. For example, if label = 'test', and a current terminal node is unnamed, it will becomes list(current_node, test = other_runner_node).

flat
logical. Whether to use the stageRunner$append method to overlay, or simply overwrite the given label. If flat = TRUE, you must supply a label. The default is flat = FALSE.